Setting journal_mode=WAL;
sqlite's default journal_mode is DELETE.
This change switches the mode from DELETE to WAL.
WAL means Write-Ahead Logging, WAL allows for
multiple threads to read from the DB as it is
being written to and vice versa.

The main advantage to this is that it allows multi user/
multi thread reads/writes without locking the whole db
table. There is a performance hit on reads according to
the docs but for now it doesn't really matter for our
use case.